Update dataset loading
Allow a list of datasets to be specified and passed to the server. This
parameter can then be used to ensure the given datasets are returned.

This will be used to restore the state when links are reloaded/shared.

The backend code to support this hasn't yet been implemented as this
isn't actually needed until with support pagination on the datasets.json
endpoint.