Refactor filter_area endpoint into explore.json
Filter params can be passed to the endpoint which applies filtering for
datasets and returns list geometries to be displayed on the map.

Examples requests:
  /explore.json?party__contains=Labour
  /explore.json?party__not_contains=Labour

Params are broken down as: <dataset>__<field_lookup>=<value>

Field lookups can be prefixed with `not_` to call `exclude` instead of
`filter`. There are many valid lookups [1] and this also supports
chaining additional lookups such as `year__gte`.

[1] https://docs.djangoproject.com/en/4.1/ref/models/querysets/#field-lookups