DEV: Add dynamic map updates
We might not want this in production as this changes the active URL and
will means extra hits to the server for filtering results but this can
be useful for development and means we don't need to click the update
button all the time.

This causes many API requests to trigger but it doesn't ensure the last
request is the one which is rendered. Rather the response which is
slowest will be rendered.