split initial geometry and filter results on explore page
Fetching all the polygons is a bit slow and once we have them we don't
need them again, and only need the list to show and any shading. This
splits the data filtering and the polygon fetching into two calls and
only uses the filtering call when the filters/shaders have changed.

Fixes #96