Update management scripts to use script/dev-command
Now that we have a generic script for running arbitrary commands either
inside or outside the container, it makes sense to use that script in
all of our other ones.

Also updated the README to reflect the fact that you can now run all
of these commands from inside or outside the container.