Add 'set to 0' to base importer
Currently, this has been implimented in the base importer. This,
in theory propogates to all commands which subclass from this.

That's said, this needs to be tested for each other importer,
partially because some should be altered so that they don't fill
out the 0s, and some don't work with this yet (e.g. CAFOD import).

Additionally, some imports should have this added if they
don't subclass from a base importer.