Make sure latest filters are applied before downloading CSV
Fixes #203.

It feels weird if we download a CSV with the latest filters applied
but don’t also update the visible table at the same time. So we need
to do both.

The weird thing is, if we call `updateState()` first, and then kick
off a `window.location.href` redirect to the explore.csv URL, the table
never updates. (Because it’s being updated asynchronously, and the
`window.location` change means the browser thinks we’ve left the page
by then, I think?)

Anyway, so instead, we set a flag that we want a redirect to the CSV
to happen after the next table update, and then kick off the update.
Nasty, but it works.

I guess ideally we’d make `updateState()` return a Promise, so we could
chain the download onto it with a `.then()` handler. But then both
`updateMap()` and `updateTable()` would need to return Promises too,
which feels like a load of faff?

I’m also aware that it’s wasteful to be requesting explore.csv TWICE
from the server (once via fetch, to update the table, and again, via
window.location.href, to download the file). Theoretically we could
replace the `window.location.href` redirect with something that writes
the response.blob() to a file, eg:

    .then(data => {
       var a = document.createElement("a")
       a.href = window.URL.createObjectURL(data)
       a.download = "explore.csv"
       a.click()
       // TODO: Probably want to revokeObjectURL() too?
    })