Make sure latest filters are applied before downloading CSV
Fixes #203.

It feels weird if we download a CSV with the latest filters applied
but don’t also update the visible table at the same time. So we need
to do both.

The weird thing is, if we called `updateState()` first, and then kicked
off a `window.location.href` redirect to the explore.csv URL, the table
never updates. (Because it’s being updated asynchronously, and the
`window.location` change means the browser thinks we’ve left the page
by then, I think?)

One way to solve this would be to make `updateState()` return a Promise,
so we could chain the `window.location.href` explore.csv redirect onto
it with a `.then()` handler. But then both `updateMap()` and
`updateTable()` would need to return Promises too, sigh.

A hackier way to solve it might have been to set a flag in the app state
saying that the user wants to download a CSV the next time the table is
updated, and then kick off `updateState()`, and have `updateTable()` do
the `window.location.href` dance _after_ it’s updated the table.

It was at this point that I realised we’re _already_ loading the CSV in
`updateTable()`, and thanks to `URL.createObjectURL()` we’re able to
create links to any old blob of data, client-side. (People typically use
this to display a client-side preview of an image selected in a file
upload input, but by assigning the URL as the href of an anchor, and
then triggering an artificial click on the anchor, you can make the
browser download the URL as a file just as easily.)

Anyway, this is also a nice way to avoid requesting the CSV _twice_ in
quick succession – it’s requested once (to update the table view), and
then we conveniently also pipe it to a file for the browser to download.

The downside is that, if we ever change `updateTable()` to use something
other than the explore.csv endpoint (to enable in-cell links for #208,
perhaps) then this will no longer work, as we’ll need to either generate
a CSV client-side, or request it separately from the server again.