Allow non-filterable datasets to be added to table as columns
`ExploreDatasetsJSON` previously filtered out all `is_filterable=False`
datasets. But we specifically _want_ people to be able to add _any_
dataset as a column on the table view.

This commit makes `ExploreDatasetsJSON` return _all_ datasets, and
we decide whether to show or hide the datasets in the modal dataset
picker based on whether they are `is_filterable`, or whether we’re
adding columns to the table view.