New homepage layout and content
- Moved search box into homepage hero.
- Visualisation of map interface, and new text content, for explore
  section on homepage.
- Homepage now accessible (with "Member sign in" prompts) to
  non-logged-in users.

We’ve wanted to rearrange the homepage for a while. Some people felt
that both CTAs (search and explore) should be visible much higher up
the page. I experimented with this, but it always felt confusing to
have two CTAs with equal prominence right at the top.

Here, we prioritise the search CTA, putting it up top in the hero, in
classic mySociety style. Everyone knows their postcode or MP, so it’s
an immediate way for them to try out the service, and see some data
they’re invested in.

This frees up extra space below to make a bigger deal out of the second
half of the site, the explore page – rather than listing the types of
data, we show a fake visualisation of two example datasets.

Fixes #144.