add exclusion attributes to DataSet model
These two fields are both used to achieve exclusion of certain countries
from datasets if the scope of the dataset doesn't extend to them, so
that they don't get filled with 0s, which would be misleading. This is
directly achieved with the exclude_countries attribute.

In addition to this, fill_blanks has been added - this functionality
has already been added (806141ee3b5ff12369da8a23b10e924ae1a8efea), but
because it's just present in the importers, it can't be referenced when
handling datasets after the initial import. Therefore, this change will
allow for greater flexibility in programmatically interacting with
datasets.