display more helpful map tooltips and table column headers
Include the filtered data in the geojson returned by explore calls so we
can then use this to include it in the tool tips.
As a result of this we're also returning data labels rather than the
internal names so this results in more useful column headers in the
table view.

Fixes #207
Part of #208