DATASET: Air quality data
The file which is used to match the 'gridcode' value to the constituency
code was created using shapefiles of constituencies re-projected onto
the same projection as the defra data (BNG 27700), and one example of
the gridcodes to x and y values as found from an arbitrary input file
for this import. QGIS was used to find where the points were within the
boundaries of the constituencies, and indeed which constituencies. To
prevent this process having to be repeated, the output file is stored
in the Drive