handle shaders for datasets without data for every area
If a dataset didn't have data for an area (e.g England only data) then
there was an error when the code tried to get the data for that area.
We skip trying to look up data for areas without a label for the opacity
data because those don't have data.

Fixes #315