use minimal CSS include on front page
Not sure this is worth it as there's a lot of bootstrap you have to
still include but this reduces the uncompressed CSS from 489 to 317k on
the front page.

Things to note are the homepage only utilities and also the removing of
the xxl breakpoint, which is universal, shaved 45k off the main css and
I don't think breaks anything.

FWIW, a big chunk of the CSS size is coming from the bootstrap
utilities/api module so if we were really optimising we could dig into
that and create our own utilities list which would probably shave off
quite a bit more.