Show MP/PPC parties on area search results page
Thanks to LIH user Marcus for suggesting this improvement!

Note: rather than enriching the person data objects with party
information inside the AreaSearchView, I’ve added a `party` method
to the `Person` model, so we can call `person.party` just like
any other person attribute. It’s not quite as conceptually pure
(since it’s looking up a `PersonData` value from elsewhere in the
database) but it saves a lot of repetition in all the views (which
I’ll refactor out in the next commit).