change to using areas many to many for person -> areas
Do this because once we get the results for the 2024 election you will
get some people who were MPs in a pre 2023 constituency and also MPs in
a 2024 constituency and we don't want to have multiple person records
for them.

This ports all the data and retains the old data in old_area for now.