Allow pinpointing of locations on Explore map
Uses the Leaflet GeoSearch plugin, and OSM Nominatim geocoder, to allow
users to place a marker on the map, at the requested location.

Handy when you know the name of a place, but not which constituency
it’s in!

Once the marker is plotted onto the map, we zoom to it. This revealed
the same Vue.js proxying issue we’ve encountered in other projects.
Namely: when you store a `Leaflet.Map` instance in the Vue.js state
(making it accessible inside your Vue methods as, say, `this.map`),
Vue _actually_ stores a wrapped version of the original instance,
to enable its clever reactive features. This “proxy” object, though,
breaks some internal Leaflet stuff, including `animateZoom` (see
https://stackoverflow.com/questions/65981712), which results in
broken zooming/panning. The recommendation from the Vue.js docs
is to “un-wrap” that proxy object when you need to work  with it,
replacing `this.map` with `toRaw(this.map)`.

I didn’t _need_ to use `toRaw()` in every place that I’ve used it here
(just inside the `geosearch/showlocation` callback would have done,
because that’s the only place we _currently_ zoom to bounds), but I
figure it’s simpler to use it _everywhere_ we refer to `this.map` than
to use it inconsitently and force future maintainers to keep making the
decision as to whether it’s needed or not.