don't return new modal dialogs if they're already defined
In theory I think vue should keep track of this but there's some sort of
interaction that happens which means it can lose track of the dialog which
means that e.g dialog.hide() does not work. To avoid this check for an
existing dialog and return that if so then return that, otherwise create
a new one.