Update import scripts to handle 2010 constituencies again
We removed 2010 WMC constituencies from the import scripts back in July,
because we figured all 2010 data had already been imported into the live
site, and we wouldn’t need to import any more.

But it’s actually useful to be able to recreate the situation of the live
site (with multiple generations of WMCs) in local development too, which
means we need the import scripts to keep working with WMCs, rather than
relying on developers just having the right data already in their local
environments.

This commit reinstates the WMC (rather than WMC23) AreaType, and updates
our mapit utility function to accept a `generation` parameter that will
easily allow us to change the MapIt generation in future again too.

With both 2010 and 2023 constituencies coming straight from MapIt now,
there is no longer any need for the import_new_constituencies command.