Allow admins to order bars in Area page bar charts
If admins set a custom `order` on a `DataType` in a `DataSet` in the
admin interface, then that order will be used when the datatypes are
rendered in the front-end, rather than just defaulting to the
alphabetical order of the datatype’s internal `name`.

This means we now have control over the order of the bars in bar charts
on the area pages.

Part of #465 (the other part being _actually_ setting these order
values, manually).