import command that uses config to get data details
This adds an `import_from_config` command that gets the details of the
imports from a JSON file. You supply the location of the file and the
name of the config you want and then it pulls in all the information
from the config.

At the moment it only works with CSV and Excel files where the data is
one row per constituency.