[WIP] Handle JSON data files in import_from_config
Largely untested!!

Enables import_from_config to create a Pandas DataFrame from a list of
dictionaries in a JSON file, in addition to rows in CSV and Excel files.

To handle complex JSON files (where the root element isn’t just a list
of the items/rows you want to import), we allow the config file to
specify a jsonpath expression, which is used to identify the bit of the
JSON object that we want to pull out.

I wrote this code, hoping to import some data from a JSON file, but then
realised the JSON file lacked constituency IDs, so would need to be
geocoded first, making this change redundant (since whatever script
does the geocoding could also output CSV instead of JSON). But maybe
this code might be useful in future..?