Improve display of AreaType description in Django admin form
The AreaType description is a full sentence. It feels odd to enter it
into a non-scrollable text `<input>` in the Django admin, rather than
a `<textarea>`.

Changing the field type from `CharField` to `TextField` would solve
this, but would introduce another migration to the code. And the 300
character limit on the CharField is still fine for our purposes.

So, instead, this commit customises the AreaType admin form, to use
a textarea widget for this field, without modifying the model.