Correctly handle DataTypes provided in Explore page URL state
It appears we had a bug, where a DataType provided in the URL state
(Eg: `/explore/?hnh_mrp_24-3_very_likely__gte=12`) would result in
a different DataType of that DataSet (perhaps just the first child
DataType in the datasets.json?) being requested from the ExploreJSON or
ExploreCSV endpoints.

`addFilter()` now plucks the correct DataType out of the DataSet
returned by `getDataset()`, if that’s what the user requested.

I am mystified as to why this doesn’t affect `addColumn()` and
`addShader()` as well, but in testing, they both seem to correctly
pull out just the requested DataType. I must be missing something.

I also fixed our FiterMixin on the server side, to correctly append
the requested DataType name to the end of the column header, when a
DataType was requested via the ExploreJSON or ExploreCSV endpoints.