fix bug restoring state when filtering a shading on same dataset/type
When restoring state we set up a bunch of anonymous functions to set up
filters and shaders which are then called once the datasets have been
fetched. For dataset/types and shaders the hash holding the promises is
keyed by the name of the dataset. This means if you have a shader and a
datatype with the same name whichever of those is second in the url
params, usually the shader, will overwrite the promise for the first.

This fixes this by appending '_shader' to the shader key.